requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
pyyaml==6.0
